Delegates
A delegate is a type that represents references to methods with a
particular parameter list and return type. When you instantiate a
delegate, you can associate its instance with any method with a
compatible signature and return type. You can invoke (or call) the
method through the delegate instance.
Delegates are used to pass methods as arguments to other methods. Event
handlers are nothing more than methods that are invoked through
delegates. You create a custom method, and a class such as a windows
control can call your method when a certain event occurs.