Create a DLL and Reference it
Visual studio is king. This is an amazing IDE. I created a class library
and then added some code in it. I saved it as its own project and then
switched back to my original solution. Within it, I added a reference to
the DLL in the bin folder of the class library solution. Badda bing!
Also, updated framework version from 4.5 to 4.5.1.