Generics
Generics let you tailor a method, class, structure, or interface to the
precise data type it acts upon. For example, instead of using the
Hashtable class, which allows keys and values to be of any type, you can
use the Dictionary<TKey, TValue> generic class and specify the type
allowed for the key and the type allowed for the value. Among the
benefits of generics are increased code reusability and type safety.